Move to separate module and add support for class methods
